refactor: decompose persistence repositories.py + decision_repo.py per-entity/aspect (#2050)
