+++ src/synthorg/persistence/sqlite/__init__.py
@@ -10,15 +10,19 @@ from synthorg.persistence.sqlite.backend import SQLitePersistenceBackend
 from synthorg.persistence.sqlite.checkpoint_repo import (
     SQLiteCheckpointRepository,
 )
+from synthorg.persistence.sqlite.cost_record_repo import (
+    SQLiteCostRecordRepository,
+)
 from synthorg.persistence.sqlite.flight_recorder_repo import (
     SQLiteFlightRecorderFrameRepository,
 )
 from synthorg.persistence.sqlite.heartbeat_repo import (
     SQLiteHeartbeatRepository,
 )
-from synthorg.persistence.sqlite.repositories import (
-    SQLiteCostRecordRepository,
+from synthorg.persistence.sqlite.message_repo import (
     SQLiteMessageRepository,
+)
+from synthorg.persistence.sqlite.task_repo import (
     SQLiteTaskRepository,
 )
