+++ tests/unit/persistence/sqlite/test_decision_repo.py
@@ -11,7 +11,7 @@ from synthorg.core.enums import DecisionOutcome
 from synthorg.core.persistence_errors import DuplicateRecordError, QueryError
 from synthorg.engine.decisions import DecisionRecord
 from synthorg.persistence.decision_protocol import DecisionRepository
-from synthorg.persistence.sqlite.decision_repo import SQLiteDecisionRepository
+from synthorg.persistence.sqlite.decision import SQLiteDecisionRepository
 from tests._shared.persistence import make_private_write_context
